// frontend/xacria-presenze/src/generated-client/model/openClosedRequestNumberResponse.ts
/**
 * API Presenze-Pro
 * API used by the Presenze-Pro thesis project’s client
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface OpenClosedRequestNumberResponse { 
    OPEN?: number;
    CLOSED?: number;
}